Avoid invalidating vendored crates
Firefox uses crates hosted on their own infrastructure with Firefox
names. Renaming that branding will invalidate the vendored crates.

So this change will try stop all of that in as minimal a way as possible,
unfortunately this still requires fine tuning firefox-on-glean /
icecat-on-glean as it's used in places outside of the obvious rust context.

Modified-by: Mark H Weaver <mhw@netris.org>
Signed-off-by: Mark H Weaver <mhw@netris.org>